### 2017-03-08 #201 @whedon
[PRE REVIEW]: Text-Fabric: Text representation for text with stand-off annotations and linguistic features
**Submitting author:** @dirkroorda (<a href="http://orcid.org/0000-0002-2680-8724">Dirk Roorda</a>)
**Repository:** <a href="https://github.com/ETCBC/text-fabric" target ="_blank">https://github.com/ETCBC/text-fabric</a>
**Version:** v2.3.6
