### 2017-04-18 #201 @acabunoc
Hi @jkahn @richardlitt @hausen @desilinguist 👋 ! Would any of you be up for reviewing this? You all seems to have python experience along with either lingustics or graphs.

Review guidelines can be found here: http://joss.theoj.org/about#reviewer_guidelines
