### 2017-04-18 #201 @jkahn
Interested, but like Nitin somewhat swamped at the moment. Perhaps a little
less swamped than Nitin.

Ping me again if you haven't found anyone by... the end of April?

--Jeremy

On Tue, Apr 18, 2017 at 1:16 PM, Nitin Madnani <notifications@github.com>
wrote:

> Unfortunately, I'm completely swamped for the next month or so. Apologies.
>
