### 2017-04-19 #201 @arfon
> Taking a quick look now, I am confused; I don't see a PDF. Is the paper.md the only submission, here?

@RichardLitt - we generally compile the PDF after review these days.
