### 2017-06-05 #201 @arfon
@47rooks - would you be willing to review this submission for JOSS?
