### 2017-06-15 #294 @47rooks
@arfon The license is Unlicense (http://unlicense.org) which is not listed on the OSI page you list but I would expect it to qualify at least from an intent standpoint. Is this ok ? Also the Github version has shifted from 2.3.6 to 2.3.7 in the time between submission and now I guess.

Regarding installation I would only mention that Python 3.6 is required at a minimum rather than just 3.
I don't see any community guidelines as such. Obviously, pinging the author via github or raising issues would work. There are no contributor guidelines per se.

Other than those minor nits it is in compliance with your guidelines.
