### 2017-06-15 #294 @arfon
> @arfon The license is Unlicense (http://unlicense.org) which is not listed on the OSI page you list but I would expect it to qualify at least from an intent standpoint. Is this ok ? Also the Github version has shifted from 2.3.6 to 2.3.7 in the time between submission and now I guess.

Actually this isn't acceptable to JOSS (see a similar recent thread [here](https://github.com/openjournals/joss/issues/248)). Probably the easiest option would be to dual-license the repository with an [OSI approved](https://opensource.org/licenses/alphabetical) license. @dirkroorda - would this be possible?
