### 2017-06-19 #294 @dirkroorda
I like to stick to the Unlicense. I find it the best expression of my intent. This piece of software is like the pond in front of my house: not part of the world of copyright. It is also simple. A dual-license is an unnecessary complication. I've seen discussions why the Unlicense is not good enough, but I have not been persuaded.
If this means that this repo cannot be reviewed for the JOSS, so be it.
Sorry for wasting your time in that case.
