### 2017-06-19 #294 @arfon
> If this means that this repo cannot be reviewed for the JOSS, so be it.

I'm afraid it does. No hard feelings 😄.

@47rooks - I'm going to withdraw the paper on @dirkroorda's behalf on the JOSS site. Many thanks for your review.
